requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pandas>=2.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
# passlib 1.7.4 + bcrypt 4.1+ rompe el hash (error 500 en /register); mantener <4.1
bcrypt>=4.0.1,<4.1.0
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.0
pydantic-settings>=2.0
psycopg2-binary>=2.9.9
slowapi>=0.1.9
